+++ myLib.py
@@ -1,7 +1,3 @@
-from imgLib import *
-from keras import backend as K
-
-
 class Cfg:
     src_dir = "./imsrc/"
     out_dir = "./imout/"
@@ -43,20 +39,3 @@ class Output:
 
     def run(self, img, operation):
         pass
-
-
-def simple_run(network, operation, output):
-    K.set_learning_phase(0)  # disables all training specific operations
-
-    net = network()
-    op = operation(net.model)
-    out = output()
-
-    # get the input img
-    img = out.load(Cfg.src_dir + Cfg.img_nm, net.preprocess_image)
-
-    # deep dream algorithm
-    gif = out.run(img, op)
-
-    # write out the dream sequence as gif
-    save_gif(Cfg.img_nm_base + "_dreaming.gif", gif, net.deprocess_image)
